Update the review page to use a single 'Update page' button
